pyproject.toml
[tool.poetry]
name = "event-finder-api"
version = "0.1.0"
description = ""
authors = ["siegerts"]
readme = "README.md"
packages = [{include = "event_finder"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
sqlmodel = "^0.0.14"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pydantic = "^2.5.0"
requests = "^2.31.0"
openai = "^1.0.0"
beautifulsoup4 = "^4.12.0"
python-dateutil = "^2.8.0"
pytest = "^7.4.0"
firecrawl-py = "^2.16.5"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
